requirements.txt
selenium==4.20.0
python-telegram-bot==13.15
